+++ src/main/kotlin/com/dingdangmaoup/resin/pura/resin/configuration/ResinXmlConfigurationStrategy.kt
@@ -1,16 +1,11 @@
-@file:Suppress("DEPRECATION")
-
 package com.dingdangmaoup.resin.pura.resin.configuration
 
 import com.dingdangmaoup.resin.pura.ResinModel
 import com.dingdangmaoup.resin.pura.resin.ResinInstallation
 import com.intellij.execution.ExecutionException
-import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.util.io.FileUtil
 import com.intellij.openapi.util.text.StringUtil
 import org.jdom.Element
-import org.jdom.JDOMException
-import org.jdom.xpath.XPath
 import java.io.File
 import java.io.InputStream
 
@@ -26,26 +21,19 @@ open class ResinXmlConfigurationStrategy(resinInstallation: ResinInstallation) :
 
     override fun getDefaultResinConfContent(): InputStream? = javaClass.getResourceAsStream(RESIN_CONF)
 
-    @Suppress("DEPRECATION")
     protected fun resolveImports(rootElement: Element, imports: MutableList<ResinConfigImport>?) {
-        try {
-            val confFolder = File(getInstallation().getResinHome(), "conf")
-            val confFolderPath = FileUtil.toSystemIndependentName(confFolder.absolutePath)
-            for (importAttrName in IMPORT_ATTRIBUTE_NAMES) {
-                val xpath =
-                    XPath.newInstance(".//*[name()='resin:import' or name()='resin:properties'][@$importAttrName]")
-                @Suppress("UNCHECKED_CAST")
-                val elements = xpath.selectNodes(rootElement) as List<Element>
-                for (element in elements) {
-                    val path = element.getAttributeValue(importAttrName)
-                    element.setAttribute(importAttrName, StringUtil.replace(path, CONF_FOLDER_VAR, confFolderPath))
-                    if (imports != null && StringUtil.equals(IMPORT_SINGLE_PATH_ATTRIBUTE, importAttrName)) {
-                        imports.add(ResinConfigImport(element))
-                    }
+
+        val confFolder = File(getInstallation().getResinHome(), "conf")
+        val confFolderPath = FileUtil.toSystemIndependentName(confFolder.absolutePath)
+        val elements = findImportElements(rootElement)
+        for (importAttrName in IMPORT_ATTRIBUTE_NAMES) {
+            for (element in elements) {
+                val path = element.getAttributeValue(importAttrName) ?: continue
+                element.setAttribute(importAttrName, StringUtil.replace(path, CONF_FOLDER_VAR, confFolderPath))
+                if (imports != null && StringUtil.equals(IMPORT_SINGLE_PATH_ATTRIBUTE, importAttrName)) {
+                    imports.add(ResinConfigImport(element))
                 }
             }
-        } catch (ex: JDOMException) {
-            LOG.info(ex)
         }
     }
 
@@ -59,10 +47,18 @@ open class ResinXmlConfigurationStrategy(resinInstallation: ResinInstallation) :
     }
 
     companion object {
-        private val LOG = Logger.getInstance(ResinXmlConfigurationStrategy::class.java)
         const val IMPORT_SINGLE_PATH_ATTRIBUTE = "path"
         private val IMPORT_ATTRIBUTE_NAMES = arrayOf(IMPORT_SINGLE_PATH_ATTRIBUTE, "fileset")
+        private val IMPORT_ELEMENT_NAMES = setOf("resin:import", "resin:properties")
         private const val CONF_FOLDER_VAR = "\${__DIR__}"
         protected const val RESIN_CONF = "resin32.xml"
+
+        internal fun findImportElements(rootElement: Element): List<Element> {
+            return rootElement.getDescendants()
+                .asSequence()
+                .filterIsInstance<Element>()
+                .filter { it.getQualifiedName() in IMPORT_ELEMENT_NAMES }
+                .toList()
+        }
     }
 }